# fix: Fixes and performance improvements in utils package