@ItalyPaleAle 2025-03-15
> @ItalyPaleAle Can you add a action for the go unit tests otehr wise there isnt a reason for thme be there, Once we have that we can merge this in.

Done!