@savely-krasovsky 2025-03-16
@ItalyPaleAle header still uses `<Avatar.Image src="/api/users/me/profile-picture.png" />`. Which leads to 500 because of failed UUID validation (`me` is not a UUID).