@ItalyPaleAle 2025-03-16
@savely-krasovsky at a quick glance, looks like [this commit](https://github.com/pocket-id/pocket-id/commit/e45d9e970d327a5120ff9fb0c8d42df8af69bb38) may be related